ai/requirements.txt
numpy==1.24.3
pandas==2.0.2
xgboost==1.7.6
scikit-learn==1.3.0
web3==6.5.0
python-dotenv==1.0.0
torch==2.0.1
transformers==4.30.2
networkx==3.1
requests==2.31.0
python-jose==3.3.0
pytz==2023.3
matplotlib==3.7.2
tqdm==4.65.0 